implemented redis cache for /offers with integration test

+++ src/main/java/pl/joboffers/JobOffers/infrastructure/security/JwtConfigurationProperties.java
@@ -13,5 +13,5 @@ public class JwtConfigurationProperties {
 
     String secret;
     String issuer;
-    Long expirationDays;
+    int expirationDays;
 }
